refactor: separate option box into a component
