+++ frontend/src/metadata/cells/metadata-cells/TextMetadataCell.svelte
@@ -2,15 +2,8 @@
 	import Button from "@smui/button";
 	import { TrailingIcon } from "@smui/chips";
 	import { Label } from "@smui/common";
-	import { tooltip } from "@svelte-plugins/tooltips";
-	import AutoComplete from "simple-svelte-autocomplete";
-	import {
-		ZenoService,
-		type FilterPredicate,
-		type ZenoColumn,
-	} from "../../../zenoservice";
-	import MatchWholeWordIcon from "./static/MatchWholeWordIcon.svelte";
-	import RegexIcon from "./static/RegexIcon.svelte";
+	import { type FilterPredicate, type ZenoColumn } from "../../../zenoservice";
+	import TextMetadataOptionBox from "./TextMetadataOptionBox.svelte";
 
 	export let col: ZenoColumn;
 	export let filterPredicates: FilterPredicate[];
@@ -19,27 +12,14 @@
 	let searchString = "";
 	let notOption = false;
 	let isRegex = false;
-	let regexValid = true;
 	let caseMatch = false;
 	let wholeWordMatch = false;
-	let refresh = 0;
-	let noResultsText = "No results";
-	let results = [];
 
 	let blur = function (ev) {
 		ev.target.blur();
 	};
-	$: {
-		regexValid = true;
-		if (isRegex) {
-			try {
-				new RegExp(searchString);
-			} catch (e) {
-				regexValid = false;
-			}
-		}
-	}
 
+	/** show the string search in chips and selection bar **/
 	function setSelection() {
 		if (!searchString) {
 			return;
@@ -69,131 +49,17 @@
 
 		searchString = "";
 	}
-
-	async function searchItems(input: string) {
-		if (isRegex) {
-			try {
-				new RegExp(input);
-				noResultsText = "No results";
-			} catch (e) {
-				noResultsText = "Invalid Regex!";
-				results = [];
-				return results;
-			}
-		}
-
-		try {
-			results = await ZenoService.filterStringMetadata({
-				column: col,
-				filterString: input,
-				isRegex: isRegex,
-				caseMatch: caseMatch,
-				wholeWordMatch: wholeWordMatch,
-			});
-			return results;
-		} catch (e) {
-			results = [];
-			return results;
-		}
-	}
-
-	function optionClick(e) {
-		if (e.currentTarget instanceof HTMLElement) {
-			let id = e.currentTarget.id;
-			if (id === "notOption") {
-				notOption = !notOption;
-			} else if (id === "caseMatch") {
-				caseMatch = !caseMatch;
-			} else if (id === "wholeWordMatch") {
-				wholeWordMatch = !wholeWordMatch;
-			} else {
-				if (isRegex) {
-					isRegex = false;
-					noResultsText = "No results";
-					regexValid = true;
-				} else {
-					isRegex = true;
-				}
-			}
-		}
-		refresh++;
-	}
 </script>
 
 <div class="container">
-	<div class="option-box" style:margin-right="10px">
-		<div
-			id="notOption"
-			class="search-option"
-			style:background={notOption ? "var(--P2)" : ""}
-			on:keydown={() => ({})}
-			on:click={optionClick}
-			use:tooltip={{
-				content: "Not",
-				theme: "zeno-tooltip",
-				autoPosition: true,
-			}}>
-			~
-		</div>
-	</div>
-	{#key refresh}
-		<AutoComplete
-			id="autoinput"
-			bind:text={searchString}
-			placeholder={"Search"}
-			{noResultsText}
-			hideArrow={true}
-			searchFunction={searchItems}
-			cleanUserText={false}
-			ignoreAccents={false}
-			localFiltering={false}
-			delay={200}>
-			<div slot="no-results" let:noResultsText>
-				<span style:color={regexValid ? "" : "red"}>{noResultsText}</span>
-			</div>
-		</AutoComplete>
-	{/key}
-	<div class="option-box">
-		<div
-			id="caseMatch"
-			class="search-option"
-			style:background={caseMatch ? "var(--P2)" : ""}
-			on:keydown={() => ({})}
-			on:click={optionClick}
-			use:tooltip={{
-				content: "Match Case",
-				theme: "zeno-tooltip",
-				autoPosition: true,
-			}}>
-			Aa
-		</div>
-		<div
-			id="wholeWordMatch"
-			class="search-option"
-			style:background={wholeWordMatch ? "var(--P2)" : ""}
-			on:keydown={() => ({})}
-			on:click={optionClick}
-			use:tooltip={{
-				content: "Match Whole Word",
-				theme: "zeno-tooltip",
-				autoPosition: true,
-			}}>
-			<svelte:component this={MatchWholeWordIcon} />
-		</div>
-		<div
-			id="typeSelection"
-			class="search-option"
-			style:background={isRegex ? "var(--P2)" : ""}
-			on:keydown={() => ({})}
-			on:click={optionClick}
-			use:tooltip={{
-				content: "Use Regular Expression",
-				theme: "zeno-tooltip",
-				autoPosition: true,
-			}}>
-			<svelte:component this={RegexIcon} />
-		</div>
-	</div>
+	<TextMetadataOptionBox
+		popup={false}
+		bind:searchString
+		bind:col
+		bind:notOption
+		bind:isRegex
+		bind:caseMatch
+		bind:wholeWordMatch />
 	<Button
 		style="margin-left: 10px; height: 32px"
 		variant="outlined"
@@ -230,29 +96,6 @@
 	.container {
 		display: flex;
 		align-items: center;
-	}
-	.option-box {
-		margin-left: 10px;
-		display: flex;
-		align-items: center;
-		border: 1px solid
-			var(--mdc-outlined-button-outline-color, rgba(0, 0, 0, 0.12));
-		border-radius: var(
-			--mdc-outlined-button-container-shape,
-			var(--mdc-shape-small, 4px)
-		);
-	}
-	.search-option {
-		display: flex;
-		align-items: center;
-		padding: 2px 3px;
-		padding-left: 6px;
-		padding-right: 6px;
-		height: 26px;
-		cursor: pointer;
-	}
-	.search-option:hover {
-		background: var(--Y1);
 	}
 	.chips {
 		display: flex;
